remove DBConf.PgSchema
The OpenDBFromDBConf function used the DBConf.PgSchema variable to execute a command setting a postgres client connection parameter when first initializing the sql.DB object. This is a bad idea as the connection parameter is only applied for that specific connection. Golang's sql.DB driver transparently uses multiple connections, and there's a chance that you can get a new connection where that parameter has not been set.

http://golang.org/pkg/database/sql/#DB

> If the database has a concept of per-connection state, such state can only be reliably observed within a transaction